+++ R/3_training_F.R
@@ -10,7 +10,7 @@ setHypers_F <- function(fMs, sOut, kerType, var.known, ls_f.known, n.starts, n.p
     U <- chol(R)
 
     # 2. estimate the a priori process variance
-    message("** Computing optimal variance...")
+    if (trace) message("** Computing optimal variance...")
     sig2 <- analyticVar_llik(U, sOut, n.tr)
 
     # 3. merge hyperparameters and return
@@ -29,13 +29,13 @@ setHypers_F <- function(fMs, sOut, kerType, var.known, ls_f.known, n.starts, n.p
     }
 
     # 3. set starting points
-    message("** Presampling...")
+    if (trace) message("** Presampling...")
     spoints <- setSPoints_F(bnds, fMs, sOut, kerType, varfun, n.starts, n.presample, nugget)
 
     # 4. Perform optimization
-    message("** Optimising hyperparameters...")
+    if (trace) message("** Optimising hyperparameters...")
     optResult <- optimHypers_F(spoints, n.starts, bnds, fMs, sOut, kerType, varfun, nugget, par.clust, trace, pbars, control.optim)
-    message("** Hyperparameters done!")
+    if (trace) message("** Hyperparameters done!")
     return(optResult)
   }
 }
@@ -94,19 +94,12 @@ setSPoints_F <- function(bnds, fMs, sOut, kerType, varfun, n.starts, n.presample
 optimHypers_F <- function(spoints, n.starts, bnds, fMs, sOut, kerType, varfun, nugget, par.clust, trace, pbars, control.optim){
   # if multistart is required then parallelize, else run single optimization
   if (n.starts == 1){
-    # if (trace) {
-      optOut <- optim(par = as.numeric(spoints), fn = negLogLik_funGp_F, method = "L-BFGS-B",
-                      lower = bnds[1,], upper = bnds[2,], control = control.optim,
-                      fMs = fMs, sOut = sOut, kerType = kerType, varfun = varfun, nugget = nugget)
-    # } else {
-    #   optOut <- quiet(optim(par = as.numeric(spoints), fn = negLogLik_funGp_F, method = "L-BFGS-B",
-    #                         lower = bnds[1,], upper = bnds[2,], control = list(trace = TRUE),
-    #                         fMs = fMs, sOut = sOut, kerType = kerType, varfun = varfun, nugget = nugget))
-    # }
-
+    optOut <- optim(par = as.numeric(spoints), fn = negLogLik_funGp_F, method = "L-BFGS-B",
+                    lower = bnds[1,], upper = bnds[2,], control = control.optim,
+                    fMs = fMs, sOut = sOut, kerType = kerType, varfun = varfun, nugget = nugget)
   } else {
     if (is.null(par.clust)) {
-      message("** Parallel backend register not found. Multistart optimizations done in sequence.")
+      if (trace) message("** Parallel backend register not found. Multistart optimizations done in sequence.")
 
       # set up progress bar
       if (pbars) {
@@ -118,15 +111,9 @@ optimHypers_F <- function(spoints, n.starts, bnds, fMs, sOut, kerType, varfun, n
       for (i in 1:n.starts) {
         modeval <- tryCatch(
           {
-            # if (trace) {
-              optim(par = as.numeric(spoints[,i]), fn = negLogLik_funGp_F, method = "L-BFGS-B",
-                    lower = bnds[1,], upper = bnds[2,], control = control.optim,
-                    fMs = fMs, sOut = sOut, kerType = kerType, varfun = varfun, nugget = nugget)
-            # } else {
-            #   quiet(optim(par = as.numeric(spoints[,i]), fn = negLogLik_funGp_F, method = "L-BFGS-B",
-            #               lower = bnds[1,], upper = bnds[2,], control = list(trace = TRUE),
-            #               fMs = fMs, sOut = sOut, kerType = kerType, varfun = varfun, nugget = nugget))
-            # }
+            optim(par = as.numeric(spoints[,i]), fn = negLogLik_funGp_F, method = "L-BFGS-B",
+                  lower = bnds[1,], upper = bnds[2,], control = control.optim,
+                  fMs = fMs, sOut = sOut, kerType = kerType, varfun = varfun, nugget = nugget)
           },
           error = function(e) e
         )
@@ -142,7 +129,7 @@ optimHypers_F <- function(spoints, n.starts, bnds, fMs, sOut, kerType, varfun, n
       if (pbars) close(pb)
 
     } else {
-      message("** Parallel backend register found. Multistart optimizations done in parallel.")
+      if (trace) message("** Parallel backend register found. Multistart optimizations done in parallel.")
 
       # register parallel backend
       registerDoFuture()
@@ -152,16 +139,10 @@ optimHypers_F <- function(spoints, n.starts, bnds, fMs, sOut, kerType, varfun, n
       with_progress({
         if (pbars) p <- progressor(along = 1:n.starts, auto_finish = FALSE)
         optOutList <- foreach(i = 1:n.starts, .errorhandling = "remove") %dopar% {
-          # if (trace) {
-            o <- optim(par = as.numeric(spoints[,i]), fn = negLogLik_funGp_F, method = "L-BFGS-B",
-                       lower = bnds[1,], upper = bnds[2,], control = control.optim,
-                       fMs = fMs, sOut = sOut, kerType = kerType, varfun = varfun, nugget = nugget)
-            ## cat("\n")
-          # } else {
-          #   o <- quiet(optim(par = as.numeric(spoints[,i]), fn = negLogLik_funGp_F, method = "L-BFGS-B",
-          #                    lower = bnds[1,], upper = bnds[2,], control = list(trace = TRUE),
-          #                    fMs = fMs, sOut = sOut, kerType = kerType, varfun = varfun, nugget = nugget))
-          # }
+          o <- optim(par = as.numeric(spoints[,i]), fn = negLogLik_funGp_F, method = "L-BFGS-B",
+                     lower = bnds[1,], upper = bnds[2,], control = control.optim,
+                     fMs = fMs, sOut = sOut, kerType = kerType, varfun = varfun, nugget = nugget)
+          ## cat("\n")
           if (pbars) p()
           return(o)
         }
@@ -178,7 +159,7 @@ optimHypers_F <- function(spoints, n.starts, bnds, fMs, sOut, kerType, varfun, n
     optOut <- optOutList[[which.min(fitvec)]]
   }
 
-  if (trace) message("The function value is the negated log-likelihood")
+  if (isTRUE(control.optim$trace)) message("The function value is the negated log-likelihood")
 
   # recovering relevant information for the estimation of the process a priori variance
   thetas_f <- optOut$par