+++ R/3_training_SF.R
@@ -10,7 +10,7 @@ setHypers_SF <- function(sMs, fMs, sOut, kerType, var.known, ls_s.known, ls_f.kn
     U <- chol(R)
 
     # 2. estimate the a priori process variance
-    message("** Computing optimal variance...")
+    if (trace) message("** Computing optimal variance...")
     sig2 <- analyticVar_llik(U, sOut, n.tr)
 
     # 3. merge hyperparameters and return
@@ -35,13 +35,13 @@ setHypers_SF <- function(sMs, fMs, sOut, kerType, var.known, ls_s.known, ls_f.kn
     }
 
     # 3. set starting points
-    message("** Presampling...")
+    if (trace) message("** Presampling...")
     spoints <- setSPoints_SF(bnds, sMs, fMs, sOut, kerType, varfun, ls_s.known, ls_f.known, n.starts, n.presample, nugget)
 
     # 4. Perform optimization
-    message("** Optimising hyperparameters...")
+    if (trace) message("** Optimising hyperparameters...")
     optResult <- optimHypers_SF(spoints, n.starts, bnds, sMs, fMs, sOut, kerType, varfun, ls_s.known, ls_f.known, nugget, par.clust, trace, pbars, control.optim)
-    message("** Hyperparameters done!")
+    if (trace) message("** Hyperparameters done!")
     return(optResult)
   }
 }
@@ -111,21 +111,13 @@ setSPoints_SF <- function(bnds, sMs, fMs, sOut, kerType, varfun, ls_s.known, ls_
 optimHypers_SF <- function(spoints, n.starts, bnds, sMs, fMs, sOut, kerType, varfun, ls_s.known, ls_f.known, nugget, par.clust, trace, pbars, control.optim){
   # if multistart is required then parallelize, else run single optimization
   if (n.starts == 1){
-    # if (trace) {
-      optOut <- optim(par = as.numeric(spoints), fn = negLogLik_funGp_SF, method = "L-BFGS-B",
-                      lower = bnds[1,], upper = bnds[2,], control = control.optim,
-                      sMs = sMs, fMs = fMs, sOut = sOut, kerType = kerType,
-                      varfun = varfun, ls_s.known = ls_s.known, ls_f.known = ls_f.known, nugget = nugget)
-    # } else {
-    #   optOut <- quiet(optim(par = as.numeric(spoints), fn = negLogLik_funGp_SF, method = "L-BFGS-B",
-    #                         lower = bnds[1,], upper = bnds[2,], control = list(trace = TRUE),
-    #                         sMs = sMs, fMs = fMs, sOut = sOut, kerType = kerType,
-    #                         varfun = varfun, ls_s.known = ls_s.known, ls_f.known = ls_f.known, nugget = nugget))
-    # }
-
+    optOut <- optim(par = as.numeric(spoints), fn = negLogLik_funGp_SF, method = "L-BFGS-B",
+                    lower = bnds[1,], upper = bnds[2,], control = control.optim,
+                    sMs = sMs, fMs = fMs, sOut = sOut, kerType = kerType,
+                    varfun = varfun, ls_s.known = ls_s.known, ls_f.known = ls_f.known, nugget = nugget)
   } else {
     if (is.null(par.clust)) {
-      message("** Parallel backend register not found. Multistart optimizations done in sequence.")
+      if (trace) message("** Parallel backend register not found. Multistart optimizations done in sequence.")
 
       # set up progress bar
       if (pbars) {
@@ -137,17 +129,10 @@ optimHypers_SF <- function(spoints, n.starts, bnds, sMs, fMs, sOut, kerType, var
       for (i in 1:n.starts) {
         modeval <- tryCatch(
           {
-            # if (trace) {
-              optim(par = as.numeric(spoints[,i]), fn = negLogLik_funGp_SF, method = "L-BFGS-B",
-                    lower = bnds[1,], upper = bnds[2,], control = control.optim,
-                    sMs = sMs, fMs = fMs, sOut = sOut, kerType = kerType,
-                    varfun = varfun, ls_s.known = ls_s.known, ls_f.known = ls_f.known, nugget = nugget)
-            # } else {
-            #   quiet(optim(par = as.numeric(spoints[,i]), fn = negLogLik_funGp_SF, method = "L-BFGS-B",
-            #               lower = bnds[1,], upper = bnds[2,], control = list(trace = TRUE),
-            #               sMs = sMs, fMs = fMs, sOut = sOut, kerType = kerType,
-            #               varfun = varfun, ls_s.known = ls_s.known, ls_f.known = ls_f.known, nugget = nugget))
-            # }
+            optim(par = as.numeric(spoints[,i]), fn = negLogLik_funGp_SF, method = "L-BFGS-B",
+                  lower = bnds[1,], upper = bnds[2,], control = control.optim,
+                  sMs = sMs, fMs = fMs, sOut = sOut, kerType = kerType,
+                  varfun = varfun, ls_s.known = ls_s.known, ls_f.known = ls_f.known, nugget = nugget)
           },
           error = function(e) e
         )
@@ -163,7 +148,7 @@ optimHypers_SF <- function(spoints, n.starts, bnds, sMs, fMs, sOut, kerType, var
       if (pbars) close(pb)
 
     } else {
-      message("** Parallel backend register found. Multistart optimizations done in parallel.")
+      if (trace) message("** Parallel backend register found. Multistart optimizations done in parallel.")
 
       # register parallel backend
       registerDoFuture()
@@ -173,18 +158,11 @@ optimHypers_SF <- function(spoints, n.starts, bnds, sMs, fMs, sOut, kerType, var
       with_progress({
         if (pbars) p <- progressor(along = 1:n.starts, auto_finish = FALSE)
         optOutList <- foreach(i = 1:n.starts, .errorhandling = "remove") %dopar% {
-          # if (trace) {
-            o <- optim(par = as.numeric(spoints[,i]), fn = negLogLik_funGp_SF, method = "L-BFGS-B",
-                       lower = bnds[1,], upper = bnds[2,], control = control.optim,
-                       sMs = sMs, fMs = fMs, sOut = sOut, kerType = kerType,
-                       varfun = varfun, ls_s.known = ls_s.known, ls_f.known = ls_f.known, nugget = nugget)
-            ## cat("\n")
-          # } else {
-          #   o <- quiet(optim(par = as.numeric(spoints[,i]), fn = negLogLik_funGp_SF, method = "L-BFGS-B",
-          #                    lower = bnds[1,], upper = bnds[2,], control = list(trace = TRUE),
-          #                    sMs = sMs, fMs = fMs, sOut = sOut, kerType = kerType,
-          #                    varfun = varfun, ls_s.known = ls_s.known, ls_f.known = ls_f.known, nugget = nugget))
-          # }
+          o <- optim(par = as.numeric(spoints[,i]), fn = negLogLik_funGp_SF, method = "L-BFGS-B",
+                     lower = bnds[1,], upper = bnds[2,], control = control.optim,
+                     sMs = sMs, fMs = fMs, sOut = sOut, kerType = kerType,
+                     varfun = varfun, ls_s.known = ls_s.known, ls_f.known = ls_f.known, nugget = nugget)
+          ## cat("\n")
           if (pbars) p()
           return(o)
         }
@@ -201,7 +179,7 @@ optimHypers_SF <- function(spoints, n.starts, bnds, sMs, fMs, sOut, kerType, var
     optOut <- optOutList[[which.min(fitvec)]]
   }
 
-  if (trace) message("The function value is the negated log-likelihood")
+  if (isTRUE(control.optim$trace)) message("The function value is the negated log-likelihood")
 
   # recovering length-scale hypers linked to scalar and functional inputs
   ds <- length(sMs)